+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 # 0.3.29-SNAPSHOT (2025-10-12)
 
 * `today's logins` intead of `today's stocks`
+- added tasks/hx-commponent
+- added tasks/hx-logins
 
 # 0.3.28